refactor load() & store() to use shared YdsResult class
