+++ src/index.ts
@@ -1,3 +1,5 @@
 export { generateIDs } from "./idgen.js";
-export { LoadResult, load } from "./load.js";
-export { StoreResult, store } from "./store.js";
+export { load } from "./load.js";
+export { store } from "./store.js";
+export { deleteElement } from "./delete.js";
+export { YdsResult } from "./result.js";
